pyproject.toml
[tool.poetry]
name = "socnet"
version = "1.0.0"
description = "SocNet"
authors = ["monosans <hsyqixco@protonmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/monosans/socnet"

[tool.poetry.dependencies]
python = "^3.8"
channels = "^3.0"
channels-redis = "^3.4"
crispy-bootstrap5 = "^0.6"
Django = { version = "~4.0", extras = ["argon2"] }
django-auto-prefetching = "^0.2"
django-cleanup = "^6.0"
django-crispy-forms = "^1.14"
django-defender = "^0.9"
django-logentry-admin = "^1.1"
django-simple-captcha = "^0.5"
django-stubs-ext = "^0.4"
djangorestframework = "^3.13"
docutils = "^0.18"
gunicorn = "^20.1"
Pillow = "^9.1"
psycopg2-binary = "^2.9"
python-decouple = "^3.6"
uvicorn = { version = "^0.17", extras = ["standard"] }

[tool.poetry.dev-dependencies]
black = "^22.3"
django-debug-toolbar = "^3.4"
django-extra-checks = "^0.11"
django-stubs = "^1.11"
djangorestframework-stubs = "^1.6"
djlint = "^1.1"
isort = "^5.10"
mypy = "^0.950"

[tool.black]
line-length = 79
skip-magic-trailing-comma = true
extend-exclude = "/migrations/"

[tool.isort]
profile = "black"
line_length = 79
extend_skip_glob = ["**/migrations/*"]

[tool.mypy]
strict = true
plugins = ["mypy_django_plugin.main"]
exclude = "migrations/"

[tool.django-stubs]
django_settings_module = "server.settings"

[tool.djlint]
blank_line_after_tag = "load,extends"
format_attribute_template_tags = true
ignore = "H006,H030,H031"
max_line_length = 80

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
